i41: Move /styles to /src/res
