+++ src/library/components/fab.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { TouchableOpacity, Text, StyleSheet } from 'react-native';
-import { typography, colors, shapes } from '../../../frontend/styles'
+import { typography, colors, shapes } from '../../res'
 import Icon from '../../res/svg'
 
 const styles = StyleSheet.create({
