+++ src/res/colors.js
