+++ src/res/index.js
@@ -1,6 +1,6 @@
 // Combines all style variables under one file: import { typography, colors } from './frontend/styles' 
 
-import * as colors from './color';
+import * as colors from './colors';
 import * as typography from './typography';
 import * as shapes from './shapes';
