+++ src/res/shapes.js
